BE: split AIO to a file
